fix(self): clear the branch's own guardrail block — placeholder fixtures, and the registry file back under the size floor
dxkit's self-guardrail has been BLOCKED on every head of this branch
(4415d0e, 7f33643, bd73d62, ee34b6b) and nobody noticed, because it is a
CI-only gate: it runs ref-based against origin/main, and the local gate
sweep never included it. Three net-new findings, all real, none of them
the reviews' fault.

**Two net-new secrets, high.** The claude-code driver test asserted that
ambient credentials are stripped and injected ones pass through, using
key-shaped literals (`ambient-key-must-not-leak`, `explicit-ci-key`) for
the fixture values. Our own scanner is right to flag a key-shaped literal
in a committed file. The fixture value carries no semantic weight here,
so both are now bracketed placeholders, the shape `isPlaceholderSecret`
already recognizes as a benign convention.

There is a real precision gap behind this and it is deliberately NOT
fixed under release pressure: `isPlaceholderSecret` does not recognise a
low-entropy hyphenated-English value as a placeholder, and fake keys in
test fixtures are a common shape. Widening it needs care in the other
direction, since a diceware passphrase has the same shape and is a real
credential. Filed for a later pass with fixtures, per the module's stated
bias toward false negatives.

**One net-new large-file, medium.** `src/managed-artifacts.ts` crossed
the 500-line floor at 502. Rather than raise the repo's threshold or
grandfather it, the legacy workspace flag fallback moves to a sibling
`managed-artifacts-detect.ts` — a coherent unit ("infer install flags
when the manifest predates them") and a different concern from the
surface registry itself. The registry file is 442 lines and still the one
source of truth: detection continues to derive the gated surfaces from
`MANAGED_SHIP_SURFACES.detectPresent`, never a second list. Re-exported
from the original module so consumers keep one import site.

Verified the way it should have been all along: `guardrail check --mode
ref-based --ref origin/main` locally — 0 blocking, verdict PASSED, 91
pairs persisted. Suite 5,029 / 331 files; typecheck, eslint at zero
warnings, prettier, arch-check, slop and cross-ecosystem all green.

Unchanged and still disclosed: the flow gate on this repo cannot evaluate
(no committed served.json), and two coverage-drift envelope notes.
